Create DrinkQueue class
Created new class DrinkQueue. Implementation of methods should be split up between
team members.

DrinkQueue class implements IQueue interface. It is unclear though if
this is correct as the offer()-method in the IQueue interface takes in
a String object as a parameter. It would make more sense if the
parameter was a Drink object as the DrinkQueue class manages Drink
objects.