Add implementation of offer, poll & remove methods
Added implementation for the methods offer, poll & remove.

Removed IQueue interface. IQueue interface implements methods for String
objects. Methods in the DrinkQueue class work with Drink objects though.
So it does not make sense to use the IQueue interface.

Still unsure if the implementation makes sense this way, as the
instructions seem to suggest to implement the IQueue interface.